+++ lib/shopify_cli/theme/dev_server.rb
@@ -11,6 +11,7 @@ require_relative "dev_server/local_assets"
 require_relative "dev_server/proxy"
 require_relative "dev_server/sse"
 require_relative "dev_server/watcher"
+require_relative "dev_server/remote_watcher"
 require_relative "dev_server/web_server"
 require_relative "dev_server/certificate_manager"
 
@@ -26,12 +27,13 @@ module ShopifyCLI
       class << self
         attr_accessor :ctx
 
-        def start(ctx, root, host: "127.0.0.1", port: 9292, poll: false, mode: ReloadMode.default)
+        def start(ctx, root, host: "127.0.0.1", port: 9292, poll: false, editor_sync: false, mode: ReloadMode.default)
           @ctx = ctx
           theme = DevelopmentTheme.find_or_create!(ctx, root: root)
           ignore_filter = IgnoreFilter.from_path(root)
-          @syncer = Syncer.new(ctx, theme: theme, ignore_filter: ignore_filter)
+          @syncer = Syncer.new(ctx, theme: theme, ignore_filter: ignore_filter, overwrite_json: !editor_sync)
           watcher = Watcher.new(ctx, theme: theme, syncer: @syncer, ignore_filter: ignore_filter, poll: poll)
+          remote_watcher = RemoteWatcher.to(theme: theme, syncer: @syncer)
 
           # Setup the middleware stack. Mimics Rack::Builder / config.ru, but in reverse order
           @app = Proxy.new(ctx, theme: theme, syncer: @syncer)
@@ -57,9 +59,17 @@ module ShopifyCLI
 
             return if stopped
 
+            preview_suffix = editor_sync ? "" : ctx.message("theme.serve.download_changes")
+            preview_message = ctx.message(
+              "theme.serve.customize_or_preview",
+              preview_suffix,
+              theme.editor_url,
+              theme.preview_url
+            )
+
             ctx.puts(ctx.message("theme.serve.serving", theme.root))
             ctx.open_url!(address)
-            ctx.puts(ctx.message("theme.serve.customize_or_preview", theme.editor_url, theme.preview_url))
+            ctx.puts(preview_message)
           end
 
           logger = if ctx.debug?
@@ -69,6 +79,7 @@ module ShopifyCLI
           end
 
           watcher.start
+          remote_watcher.start if editor_sync
           WebServer.run(
             @app,
             BindAddress: host,
@@ -76,6 +87,7 @@ module ShopifyCLI
             Logger: logger,
             AccessLog: [],
           )
+          remote_watcher.stop if editor_sync
           watcher.stop
 
         rescue ShopifyCLI::API::APIRequestForbiddenError,