+++ lib/shopify_cli/theme/dev_server/local_assets.rb
@@ -71,7 +71,7 @@ module ShopifyCLI
 
         def replace_asset_urls(body)
           replaced_body = body.join.gsub(ASSET_REGEX) do |match|
-            path = Pathname.new(Regexp.last_match[1])
+            path = Regexp.last_match[1]
             if @theme.static_asset_paths.include?(path)
               "/#{path}"
             else